+++ tests/conftest.py
@@ -13,14 +13,6 @@ from sklearn.utils import Bunch
 from pydvl.parallel import available_cpus
 from pydvl.utils import Dataset, MemcachedClientConfig
 from tests.cache import CloudPickleCache
-from tests.tolerate import (
-    TolerateErrorFixture,
-    TolerateErrorsSession,
-    wrap_pytest_function,
-)
-
-if TYPE_CHECKING:
-    from _pytest.terminal import TerminalReporter
 
 
 def pytest_addoption(parser):
@@ -34,19 +26,6 @@ def pytest_addoption(parser):
         "--slow-tests",
         action="store_true",
         help="Run tests marked as slow using the @slow marker",
-    )
-    group = parser.getgroup("tolerate")
-    group.addoption(
-        "--tolerate-verbose",
-        action="store_true",
-        default=False,
-        help="Dump diagnostic and progress information.",
-    )
-    group.addoption(
-        "--tolerate-quiet",
-        action="store_true",
-        default=False,
-        help="Disable reporting. Verbose mode takes precedence.",
     )
 
 
@@ -202,16 +181,10 @@ def pytest_xdist_auto_num_workers(config) -> Optional[int]:
 
 
 ################################################################################
-# Tolerate Errors and CloudPickleCache Plugins
+# CloudPickleCache Plugins
 
 
 def pytest_configure(config: "Config"):
-    config.addinivalue_line(
-        "markers",
-        "tolerate: mark a test to swallow errors up to a certain threshold. "
-        "Use to test (ε,δ)-approximations.",
-    )
-    config._tolerate_session = TolerateErrorsSession(config)
     config.cloud_pickle_cache = CloudPickleCache.for_config(config, _ispytest=True)
 
     config.addinivalue_line(
@@ -233,28 +206,3 @@ def pytest_runtest_setup(item: pytest.Item):
     if marker:
         if not item.config.getoption("--slow-tests"):
             pytest.skip("slow test")
-
-
-@pytest.fixture(scope="function")
-def tolerate(request: pytest.FixtureRequest):
-    fixture = TolerateErrorFixture(request.node)
-    return fixture
-
-
-@pytest.hookimpl(hookwrapper=True)
-def pytest_runtest_call(item: pytest.Function):
-    marker = item.get_closest_marker("tolerate")
-    has_fixture = hasattr(item, "funcargs") and isinstance(
-        item.funcargs.get("tolerate"), TolerateErrorFixture
-    )
-    if marker:
-        if not has_fixture:
-            wrap_pytest_function(item)
-    yield
-
-
-def pytest_terminal_summary(
-    terminalreporter: "TerminalReporter", exitstatus: int, config: "Config"
-):
-    tolerate_session = terminalreporter.config._tolerate_session
-    tolerate_session.display(terminalreporter)